Roughly the first half of my solutions manual.
So far, I've worked on chapters 1, 2, 3, 4, 5, and 6 of Axler.  Chapters 1 and 4 are completed, 2 is nearly complete, 3 is mostly complete, and 5 and 6 are both slightly less than half complete.  That said, the work that I have done on these problems still comprises approximately half of the over 500 problems in the book.